Tic-Tac-Toe game that can be run on the console of BlueJ
The game was created with an already made framework. The implementation
of the winning conditions and players' move was added by me. I also did
my best to prevent error occurs in the game if the player made an
incorrect move. But i did not do anything to incorrect input that is in
form of string. That could be prevent when I implement it using GUI